requirements.txt
# AlgoProject - Complete Trading Platform Requirements
# Fixed version - removes problematic dependencies

# ========================================================================
# CRYPTO TRADING LIBRARIES
# ========================================================================
ccxt>=4.0.0              # Cryptocurrency exchange library (100+ exchanges)
websocket-client>=1.8.0  # WebSocket connections for live crypto data

# ========================================================================  
# STOCK TRADING LIBRARIES (Alternative data sources)
# ========================================================================
yfinance>=0.2.0          # Yahoo Finance data (works reliably)
# fyers-apiv3>=3.0.0     # Disabled: Fyers API for Indian stock market
# tvdatafeed>=1.4.0      # Disabled: Package not available
# nsepython>=2.10        # Optional: NSE data and utilities
# backtrader>=1.9.0      # Optional: Backtesting framework

# ========================================================================
# CORE DATA & ANALYSIS LIBRARIES  
# ========================================================================
pandas>=1.3.0            # Data manipulation and analysis
numpy>=1.21.0            # Numerical computing
requests>=2.31.0         # HTTP requests for API calls
scipy>=1.10.0            # Statistical functions
scikit-learn>=1.3.0      # Machine learning utilities

# ========================================================================
# CONFIGURATION & UTILITIES
# ========================================================================
pyyaml>=6.0              # YAML configuration files
python-dotenv>=1.0.0     # Environment variables management
tabulate>=0.9.0          # Beautiful table formatting
pytz>=2023.3             # Timezone handling for global markets
joblib>=1.3.0            # Parallel processing and caching

# ========================================================================
# VISUALIZATION & DISPLAY
# ========================================================================
matplotlib>=3.5.0        # Plotting and charts
plotly>=5.0.0            # Interactive charts
colorama>=0.4.6          # Colored terminal output
rich>=13.0.0             # Rich text and beautiful formatting

# ========================================================================
# TECHNICAL ANALYSIS
# ========================================================================
ta>=0.10.0               # Technical Analysis library (pure Python - WORKING)
# pandas-ta>=0.3.0       # Optional: Enhanced technical analysis
# ta-lib>=0.4.0          # Disabled: Requires Visual C++ Build Tools

# ========================================================================
# TESTING FRAMEWORK DEPENDENCIES
# ========================================================================
networkx>=2.8.0          # Graph analysis for dependency checking
psutil>=5.9.0            # System and process utilities

# ========================================================================
# DEVELOPMENT & TESTING (Optional)
# ========================================================================
# pytest>=7.0.0          # Testing framework
# jupyter>=1.0.0         # Jupyter notebooks for analysis

# ========================================================================
# NOTES:
# ========================================================================
# - Removed problematic dependencies (ta-lib, tvdatafeed, fyers-apiv3)
# - Focus on crypto trading (fully functional)
# - Alternative data sources for stocks (yfinance)
# - All dependencies tested and working on Windows
